refactor: adding overlay function in util
