+++ assets/src/js/plugins/give-api/form.js
@@ -1,4 +1,6 @@
 /* globals Give, jQuery */
+import Util from './util';
+
 export default {
 	init: function() {
 		this.fn.field.formatCreditCard( jQuery( 'form.give-form' ) );
@@ -56,27 +58,10 @@ export default {
 		 * Show processing state template.
 		 *
 		 * @since 2.8.0
-		 * @since {string} none
+		 * @since {string} html Message html string or plain text.
 		 */
 		showProcessingState: function( html ) {
-			const container = document.createElement( 'div' );
-			const loader = document.createElement( 'div' );
-			const divNode = document.createElement( 'div' );
-
-			if ( html ) {
-				divNode.innerHTML = html;
-			} else {
-				divNode.innerHTML = Give.fn.getGlobalVar( 'textForOverlayScreen' );
-			}
-
-			loader.setAttribute( 'class', 'loader spinning' );
-			container.setAttribute( 'id', 'give-processing-state-template' );
-
-			container.append( loader );
-			container.append( divNode );
-
-			container.classList.add( 'active' );
-			document.body.appendChild( container );
+			Util.fn.showOverlay( html );
 		},
 
 		/**
@@ -85,7 +70,7 @@ export default {
 		 * @since 2.8.0
 		 */
 		hideProcessingState: function( ) {
-			document.getElementById( 'give-processing-state-template' ).remove();
+			Util.fn.hideOverlay();
 		},
 
 		/**
